adicionando filtro

+++ app/Http/Controllers/AppointmentController.php
@@ -6,7 +6,8 @@ use App\Http\Requests\StoreAppointmentRequest;
 use App\Http\Requests\UpdateAppointmentRequest;
 use App\Models\Appointment;
 use App\Repositories\AppointmentRepository;
-
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Http\Request;
 
 class AppointmentController extends Controller
 {
@@ -18,11 +19,31 @@ class AppointmentController extends Controller
    /**
      * Display a listing of the resource.
      */
-    public function index()
-    {
-        //
-        $appointment = $this->appointmentRepository->with(['user']);
-        return response()->json($appointment, 200);
+    public function index(Request $request)
+    {        
+        try {
+            $user = auth()->user();
+            $filters = [];
+
+            $type_animal = $request->query('tipo');
+            $date = $request->query('data');
+
+
+            if ($user->user_type == 'D')
+                $filters = ['user_id' => $user->id];
+            
+            if ($type_animal)
+                $filters['animal_type'] = $type_animal;
+            
+            if ($date)
+                $filters['appointment_date'] = $date;
+
+            $appointment = $this->appointmentRepository->with(['user'],$filters);
+
+            return response()->json($appointment, 200);
+        } catch (\Throwable $th) {
+            return response()->json([ 'message' => $th ] , 500);
+        }
     }
 
     /**
@@ -31,9 +52,14 @@ class AppointmentController extends Controller
     public function store(StoreAppointmentRequest $request)
     {
         //
-        $appointment = $request->validated();
-        $appointment = $this->appointmentRepository->create($appointment);
-        return response()->json($appointment, 201);
+        
+        try {
+            $appointment = $request->validated();
+            $appointment = $this->appointmentRepository->create($appointment);
+            return response()->json($appointment, 201);
+        } catch (\Throwable $th) {
+            return response()->json([ 'message' => $th ] , 500);
+        }
     }
 
     /**
@@ -42,8 +68,12 @@ class AppointmentController extends Controller
     public function show($id)
     {
         //
-        $appointment = $this->appointmentRepository->find($id, ['user']);;
-        return response()->json($appointment, 200);
+        try {
+            $appointment = $this->appointmentRepository->find($id, ['user']);;
+            return response()->json($appointment, 200);
+        } catch (\Throwable $th) {
+            return response()->json([ 'message' => $th ] , 500);
+        }
     }
 
     /**
@@ -52,9 +82,14 @@ class AppointmentController extends Controller
      public function update(UpdateAppointmentRequest $request, $id)
     {
         //
-        $appointment = $request->validated();
-        $appointment = $this->appointmentRepository->update($id, $appointment);
-        return response()->json($appointment, 200);
+        try {
+            $appointment = $request->validated();
+            $appointment = $this->appointmentRepository->update($id, $appointment);
+            return response()->json($appointment, 200);
+        } catch (\Throwable $th) {
+            return response()->json([ 'message' => $th ] , 500);
+        }
+        
     }
 
     /**
@@ -62,7 +97,15 @@ class AppointmentController extends Controller
      */
     public function destroy($id)
     {
-        $appointment = $this->appointmentRepository->delete($id);
-        return response()->json([ 'message' => 'deletado' ] , 200);
+        try {
+            $user = auth()->user();
+            if($user->user_type == 'D')
+                return response()->json([ 'message' => 'Voce não tem permissão para deletar' ] , 401);
+            $appointment = $this->appointmentRepository->delete($id);
+            return response()->json([ 'message' => 'deletado' ] , 200);
+        } catch (\Throwable $th) {
+             return response()->json([ 'message' => $th ] , 500);
+        }
+        
     }
 }
